+++ dam/__init__.py
@@ -1,2 +1,2 @@
 """Docker Automation Manager"""
-__version__ = "0.6.0"
+__version__ = "0.7.0"